scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
transformers>=4.30.0
requests>=2.31.0
python-dotenv>=1.0.0
sentencepiece>=0.1.99
accelerate>=0.20.0
tqdm>=4.64.0
psutil>=5.9.0
pynvml>=11.4.1
safetensors>=0.3.1
